requirements.txt
pytest>=8.0.0
selenium>=4.15.0
Pillow>=10.0.0
ImageHash>=4.3.0
allure-pytest>=2.13.0
